Working to fix up longrange in Win32 version. This seemed to fix Lance's issues for now.
git-svn-id: svn+ssh://cmms01.ace-net.ca/users/cmms/jmercer/programming/repos/maglua@209 01fa2656-f03c-45d1-8fba-b86aa1ed0177

+++ src/modules/common/encode/luamigrate.cpp
@@ -27,7 +27,7 @@ static int lexportwriter(lua_State *L, const void* chunk, size_t size, void* dat
 {
 	(void)L;
 	buffer* b = (buffer*)data;
-	encodeBuffer(chunk, size, b);
+	encodeBuffer(chunk, (const int)size, b);
 	return 0;
 }
